Beginnings of line chart before backend update
