+++ housing_enjoyers_website/package.json
@@ -8,11 +8,13 @@
 		"@mui/base": "^5.0.0-beta.61",
 		"@mui/icons-material": "^6.1.5",
 		"@mui/material": "^6.1.5",
+		"@mui/x-date-pickers": "^7.22.1",
 		"@testing-library/jest-dom": "^5.17.0",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^13.5.0",
 		"axios": "^1.7.7",
 		"chart.js": "^4.4.6",
+		"dayjs": "^1.11.13",
 		"react": "^18.3.1",
 		"react-chartjs-2": "^5.2.0",
 		"react-dom": "^18.3.1",
