+++ housing_enjoyers_website/src/App.js
@@ -3,6 +3,7 @@ import React from 'react';
 import { Grid, Container, Paper } from '@mui/material';
 import { BrowserRouter as Router, Route, Routes, Link, BrowserRouter } from 'react-router-dom';
 import Affordability from './pages/Affordability.js';
+import Prediction from './pages/Prediction.js';
 import Home from './pages/Home.js';
 
 
@@ -15,6 +16,7 @@ function App() {
     <Routes>
       <Route path='/' element={<Home/>} />
       <Route path='/Affordability' element={<Affordability/>} />
+      <Route path='/Prediction' element={<Prediction/>} />
 
     </Routes>
