+++ housing_enjoyers_website/src/pages/Prediction.js
@@ -1,21 +1,26 @@
-import ResponsiveAppBar from '../components/navbar.js';
-import createPieChart from '../services/createPieChart.js';
+// Libraries
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
 import { 
     Grid2 as Grid, 
     Container, TextField, Box, Button, Typography 
 } from '@mui/material';
-import axios from 'axios'
-import React, { useEffect, useState } from 'react';
-import BedBathAffordability from '../components/BedBathAffordability.js'
+import DatePicker from '@mui/x-date-pickers/DatePicker'
+// Local Files
+import ResponsiveAppBar from '../components/navbar';
+import BedBathAffordability from '../components/BedBathAffordability';
+import createPieChart from '../services/createPieChart';
+import createLineChart from '../services/createLineChart';
 
 
-function Affordability() {
+function Prediction() {
 
     const [pieChartData, setPieChartData] = useState({});
     const [borrow, setBorrow] = useState('');
     const [currentPieChart, setCurrentPieChart] = useState('');
     const [barChartData, setBarChartData] = useState('');
     const [lineChartData, setLineChartData] = useState('');
+    const [currentLineChart, setCurrentLineChart] = useState('');
 
     // useeffect runs the code below on first render to get the default pie chart and bar charts
     useEffect(() => {
@@ -88,62 +93,51 @@ function Affordability() {
     return (
     <div className="App">
         <ResponsiveAppBar/>
-
         <h1>React Line Chart with Chart.js</h1>
+
         <Container maxWidth="lg" style={{ marginTop: '200px'}} xs={{ display: 'flex' }}>
-        {/* Grid Container to center the charts */}
-        <Grid container spacing={4}>
-        
+            {/* Grid Container to center the charts */}
+            <Grid container spacing={4}>
+                <Grid container columns={2} rowSpacing={6} columnSpacing={{ xs: 1, sm: 2, md: 3 }}>
 
-        <Grid container columns={2} rowSpacing={1} columnSpacing={{ xs: 1, sm: 2, md: 3 }}>
+                    <Grid container size={2} rowSpacing={1}>
+                        <Grid container size={1}>
+                            <Typography className='text_box' align='justify' xs={{border: '2px solid #1976d2' , borderRadius: '16px'}}>
+                                The median weekly income of Australia is $1,500 in 2023. If that person works 50 weeks of the year
+                                then that comes to a total of $75,000 per year. Assuming monthly expenses of 1,000 per month, 
+                                this particular person can borrow a maximum of $394,000 to pay for their property. 
+                                This is what the pie chart to the left shows the affordability of such properties from 2017 to 2018 according 
+                                these statistics.
+                            </Typography>
+                        </Grid>
 
-        <Grid size={1}>
-            <div>{currentPieChart}</div>
-        </Grid>
+                        <Grid container size={1} direction={'column'}>
+                            <Typography align='center' style={{ paddingTop: '20px'}}>
+                                When are you looking to buy a house? Use the date pickers below!
+                            </Typography>
 
-        <Grid container size={1} direction="column">
-            <Typography className='text_box' align='justify' xs={{border: '2px solid #1976d2' , borderRadius: '16px'}}>
-                The median weekly income of Australia is $1,500 in 2023. If that person works 50 weeks of the year
-                then that comes to a total of $75,000 per year. Assuming monthly expenses of 1,000 per month, 
-                this particular person can borrow a maximum of $394,000 to pay for their property. 
-                This is what the pie chart to the left shows the affordability of such properties from 2017 to 2018 according 
-                these statistics.
-            </Typography>
+                            <Box style={{padding: '2em 4em 0em'}} >
+                                <TextField type="number" id="outlined-basic" label="Maximum borrowing amount ($)" variant="outlined" placeholder="75000" fullWidth
+                                    onChange={ (event) => { setBorrow(parseInt(event.target.value)); } }
+                                />
+                                <DatePicker label={'Year'} views={['year']} />
+                            </Box>
+                            <Grid container spacing={10} columns={2} style={{ paddingTop: '20px' }}>
+                                <Grid size={2}>
+                                    <Button variant="outlined" onClick={() => updatePieChart(borrow)}>Calculate Affordability</Button>
+                                </Grid>
+                            </Grid>
+                        </Grid>
+                    </Grid>
 
-
-            <Typography align='center' style={{ paddingTop: '20px'}}>
-                What are your affordability options looking like? Put it in the text box below!
-            </Typography>
-
-            <Box style={{padding: '3em 6em 0em'}} >
-                <TextField type="number" id="outlined-basic" label="Maximum borrowing amount ($)" variant="outlined" placeholder="75000" fullWidth
-                    onChange={ (event) => { setBorrow(parseInt(event.target.value)); } }
-                />
-            </Box>
-            <Grid container spacing={10} columns={2} style={{ paddingTop: '50px' }}>
-                <Grid size={2}>
-                    <Button variant="outlined" onClick={() => updatePieChart(borrow)}>Calculate Affordability</Button>
+                    <Grid size={2}>
+                        <div>{currentLineChart}</div>
+                    </Grid>
                 </Grid>
             </Grid>
-        </Grid>
-        </Grid>
-
-        </Grid>
-    
-       
         </Container>
-
-        <BedBathAffordability/>
-   
-   
-
     </div>
-
-    
-      
-
     );
-
 }
 
-export default Affordability;
+export default Prediction;
